+++ test/integration/tests/relative-script-snapshots/files/subdir/script.hs
@@ -1,0 +1,6 @@
+#!/usr/bin/env stack
+-- stack --resolver snapshot.yaml script
+import Acme.Missiles
+
+main :: IO ()
+main = launchMissiles